# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Облік Майна API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/PATCH/documents/realize": {
      "patch": {
        "summary": "Realize documents",
        "description": "Realize created earlier documents",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "id": "integer",
                  "operation": "string"
                }
              },
              "example": [
                {
                  "id": 1,
                  "operation": "ADD_SUPPLY"
                },
                {
                  "id": 2,
                  "operation": "ADD_SUPPLY"
                },
                {
                  "id": 5,
                  "operation": "ADD_SUPPLY"
                },
                {
                  "id": 1,
                  "operation": "PIN_MAIN_ASSET"
                }
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/POST/document/add-supply/save-realize": {
      "post": {
        "summary": "Save and realize document",
        "description": "Save and realize document",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/api-post-document-addSupply-saveRealize/requestBody.yaml#/DocumentAddSupply"
              },
              "examples": {
                "default": {
                  "$ref": "./schemas/api-post-document-addSupply-saveRealize/requestBody-example.yaml"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/api-post-document-addSupply-saveRealize/response.yaml#/DocumentAddSupply"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/POST/document/pin-ma/save": {
      "post": {
        "summary": "Save document",
        "description": "Save document",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/api-post-document-pinMainAsset-save/requestBody.yaml#/DocumentPinMainAsset"
              },
              "examples": {
                "default": {
                  "$ref": "./schemas/api-post-document-pinMainAsset-save/requestBody-example.yaml"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/api-post-document-pinMainAsset-save/response.yaml#/DocumentPinMainAsset"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}